adding getVersion tests
