+++ tests/methods/getVersion/index.js
@@ -1,0 +1,5 @@
+require('../../util');
+describe('getVersion', function(){
+	require('./instance/index');
+	require('./static/index');
+});
