+++ tests/methods/getVersion/instance/index.js
@@ -1,0 +1,6 @@
+require('../../../util');
+describe('get version using instance method', function(){
+	require('./date');
+	require('./version');
+});
+
