+++ tests/methods/getVersion/static/index.js
@@ -1,0 +1,6 @@
+require('../../../util');
+describe('get version using static method', function(){
+	require('./date');
+	require('./version');
+});
+
